# attr.asdict(recurse=True) now can handle nested lists