--- a/changelog.d/345.change.rst
+++ b/changelog.d/345.change.rst
@@ -1,0 +1,1 @@
+``attr.asdict(recurse=True)`` now can handle nested lists