Right now if field contains nested lists it will not properly encoded to dict:
```python
>>> @attr.s
... class Foo():
...     a = attr.ib()
...
>>> a = Foo(a=[[Foo(a=1)]])
>>> attr.asdict(a)
{'a': [[Foo(a=1)]]}
```

I added support for this case (only for lists and only for `asdict` function) and also updated testing strategy