@Tinche 2018-02-24
Hi,

I see two problems with this solution.

The first is we're making a new function every asdict invocation, which slows things down. The second is, it doesn't solve the problem fully. Consider this:

```
>>> b = Foo([{1: Foo(1)}])
>>> asdict(b)
{'a': [{1: Foo(a=1)}]}
```

So if we're gonna solve it, let's solve it right. I've pushed a branch with what I think is a more correct solution: https://github.com/python-attrs/attrs/tree/feature/recursive-asdict. It makes asdict in general a little slower, but it's faster than this solution.

```
Before:
> pyperf timeit -g -s "from attr import asdict; from random import Random; from tests.utils import nested_classes, simple_classes; r = Random(1); C = simple_classes().example(r); c=C()" "asdict(c)"
Mean +- std dev: 2.43 us +- 0.09 us

After:
> pyperf timeit -g -s "from attr import asdict; from random import Random; from tests.utils import nested_classes, simple_classes; r = Random(1); C = simple_classes().example(r); c=C()" "asdict(c)"
Mean +- std dev: 2.51 us +- 0.10 us

This branch:
> pyperf timeit -g -s "from attr import asdict; from random import Random; from tests.utils import nested_classes, simple_classes; r = Random(1); C = simple_classes().example(r); c=C()" "asdict(c)"
Mean +- std dev: 2.95 us +- 0.10 us

```

Please add the tests to that branch and we can get this merged.