@hynek 2018-06-17
Fixed in #395 – thank you for the inspiration.